+++ packages/amplify-codegen-e2e-core/src/index.ts
@@ -3,6 +3,8 @@ import * as path from 'path';
 import * as fs from 'fs-extra';
 import { spawnSync, execSync } from 'child_process';
 import { v4 as uuid } from 'uuid';
+import * as ini from 'ini';
+import { pathManager } from '@aws-amplify/amplify-cli-core';
 
 export * from './configure/';
 export * from './init/';
@@ -29,7 +31,14 @@ export function getCLIPath(testingWithLatestCodebase = false) {
 }
 
 export function isCI(): boolean {
-  return process.env.CI && process.env.CIRCLECI ? true : false;
+  return process.env.CI && (process.env.CODEBUILD) ? true : false;
+}
+
+export function injectSessionToken(profileName: string) {
+  const credentialsContents = ini.parse(fs.readFileSync(pathManager.getAWSCredentialsFilePath()).toString());
+  credentialsContents[profileName] = credentialsContents[profileName] || {};
+  credentialsContents[profileName].aws_session_token = process.env.AWS_SESSION_TOKEN;
+  fs.writeFileSync(pathManager.getAWSCredentialsFilePath(), ini.stringify(credentialsContents));
 }
 
 export function npmInstall(cwd: string) {