+++ packages/amplify-codegen-e2e-tests/src/__tests__/build-app-swift.test.ts
@@ -1,13 +1,11 @@
 import {
   initProjectWithQuickstart,
   DEFAULT_IOS_CONFIG,
-  addApiWithBlankSchemaAndConflictDetection,
   updateApiSchemaWithText,
   generateModels,
-  swiftBuild,
 } from '@aws-amplify/amplify-codegen-e2e-core';
 const { schemas } = require('@aws-amplify/graphql-schema-test-library');
-import { existsSync, writeFileSync, readdirSync, rmSync, readFileSync } from 'fs';
+import { writeFileSync, readdirSync, readFileSync } from 'fs';
 import path from 'path';
 
 const skip = new Set([
@@ -30,31 +28,25 @@ describe('build app - Swift', () => {
   });
 
   afterEach(async () => {
-    await rmSync(path.join(projectRoot, 'amplify', 'generated', 'models'), { recursive: true, force: true });
     writeFileSync(path.join(projectRoot, 'swift.xcodeproj', 'project.pbxproj'), projectPBXProjCache);
   });
 
   Object.entries(schemas).forEach(([schemaName, schema]) => {
     // @ts-ignore
     const testName = `builds with ${schemaName}: ${schema.description}`;
+    const schemaFolderName = schemaName.replace(/[^a-zA-Z0-9]/g, '');
+    const outputDir = path.join(projectRoot, 'amplify', 'generated', 'models', schemaFolderName);
     const testFunction = async () => {
       // @ts-ignore
       const schemaText = `input AMPLIFY { globalAuthRule: AuthRule = { allow: public } }\n${schema.sdl}`;
-      console.log(schemaText); // log so that circleci does not timeout
+      console.log(schemaText); // log so that ci does not timeout
       updateApiSchemaWithText(projectRoot, 'amplifyDatasource', schemaText);
-      await generateModels(projectRoot);
-      await swiftBuild(projectRoot, { ...config, scheme: 'swift' });
+      await generateModels(projectRoot, outputDir);
     };
     if (skip.has(schemaName)) {
       it.skip(testName, testFunction);
     } else {
       it(testName, testFunction);
     }
-  });
-
-  it('fails build with syntax error', async () => {
-    await generateModels(projectRoot);
-    await writeFileSync(path.join(projectRoot, 'amplify', 'generated', 'models', 'AmplifyModels.swift'), 'foo\nbar');
-    await expect(swiftBuild(projectRoot, { ...config })).rejects.toThrowError();
   });
 });